LayoutTests/platform/ios-sim-deprecated/fast/dom/zoom-scroll-page-test-expected.txt

This tests that window.scrollBy respects the page's zoom and scale factor when scrolling.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing at zoom -1 and scale 1.2
FAIL window.pageYOffset should be within 3 of 231. Was 220.
Testing at zoom -3 and scale 0.75
FAIL window.pageYOffset should be within 3 of 231. Was 219.
Testing at zoom -3 and scale 1
FAIL window.pageYOffset should be within 3 of 231. Was 219.
Testing at zoom -3 and scale 1.2
FAIL window.pageYOffset should be within 3 of 231. Was 219.
Testing at zoom -3 and scale 1.6
FAIL window.pageYOffset should be within 3 of 231. Was 219.
Testing at zoom -2 and scale 0.75
FAIL window.pageYOffset should be within 3 of 231. Was 220.
Testing at zoom -2 and scale 1
FAIL window.pageYOffset should be within 3 of 231. Was 220.
Testing at zoom -2 and scale 1.2
FAIL window.pageYOffset should be within 3 of 231. Was 220.
Testing at zoom -2 and scale 1.6
FAIL window.pageYOffset should be within 3 of 231. Was 220.
Testing at zoom -1 and scale 0.75
FAIL window.pageYOffset should be within 3 of 231. Was 220.
Testing at zoom -1 and scale 1
FAIL window.pageYOffset should be within 3 of 231. Was 220.
Testing at zoom -1 and scale 1.2
FAIL window.pageYOffset should be within 3 of 231. Was 220.
Testing at zoom -1 and scale 1.6
FAIL window.pageYOffset should be within 3 of 231. Was 220.
Testing at zoom 0 and scale 0.75
FAIL window.pageYOffset should be within 3 of 231. Was 221.
Testing at zoom 0 and scale 1
FAIL window.pageYOffset should be within 3 of 231. Was 221.
Testing at zoom 0 and scale 1.2
FAIL window.pageYOffset should be within 3 of 231. Was 221.
Testing at zoom 0 and scale 1.6
FAIL window.pageYOffset should be within 3 of 231. Was 221.
Testing at zoom 1 and scale 0.75
FAIL window.pageYOffset should be within 3 of 231. Was 220.
Testing at zoom 1 and scale 1
FAIL window.pageYOffset should be within 3 of 231. Was 220.
Testing at zoom 1 and scale 1.2
FAIL window.pageYOffset should be within 3 of 231. Was 220.
Testing at zoom 1 and scale 1.6
FAIL window.pageYOffset should be within 3 of 231. Was 220.
Testing at zoom 3 and scale 0.75
FAIL window.pageYOffset should be within 3 of 231. Was 220.
Testing at zoom 3 and scale 1
FAIL window.pageYOffset should be within 3 of 231. Was 220.
Testing at zoom 3 and scale 1.2
FAIL window.pageYOffset should be within 3 of 231. Was 220.
Testing at zoom 3 and scale 1.6
FAIL window.pageYOffset should be within 3 of 231. Was 220.
Testing at zoom 5 and scale 0.75
FAIL window.pageYOffset should be within 3 of 231. Was 220.
Testing at zoom 5 and scale 1
FAIL window.pageYOffset should be within 3 of 231. Was 220.
Testing at zoom 5 and scale 1.2
FAIL window.pageYOffset should be within 3 of 231. Was 220.
Testing at zoom 5 and scale 1.6
FAIL window.pageYOffset should be within 3 of 231. Was 220.
PASS successfullyParsed is true

TEST COMPLETE
